Убрал баг с дублирование списка покупок в другие списки

+++ app/src/main/java/acr/appcradle/shoppinglist/ui/screens/lists_all/nav/ListsAllNav.kt
@@ -14,7 +14,7 @@ fun NavController.navigateListsAll() = navigate(route = ListsAllRoute)
 
 fun NavGraphBuilder.listsAll(
     createNewListClick: () -> Unit,
-    onListClick: () -> Unit,
+    onListClick: (Long) -> Unit,
     onThemeChange: (ThemeOption) -> Unit
 ) {
     composable<ListsAllRoute> {
